ui: Support progress bar with unknown size
When downloading an image, we don't know the image size until we perform
a GET request for entire image, and get the content-length. Another
option is to get image extents, and sum the length of the extents.

In both cases we need to create the progress bar when the size is known,
and update the size later. Change the ui.ProgressBar() to support this
usage.

When size is not set, the progress part will show:

[ ------- ] 0.00 GiB  0.12 seconds  0.00 MiB/s

Once we set the size, the progress will be updated normally.

Change-Id: Ifcf5c213002cab4fcfa5d55ae0358e1e6ee94dbe
Signed-off-by: Nir Soffer <nsoffer@redhat.com>