client: Accept ProgressBar instead of update callable
Accepting an update callable is nice, making it easy to integrate with
other implementations. But we cannot use it for download, since in
download we don't know the download size before we connect to the
server, so we need to pass the progress bar itself, so the download code
can set the progress size at the start of the download.

Change-Id: I65f39a2e85778f9ee35fd4710ed5e64927e8d69f
Signed-off-by: Nir Soffer <nsoffer@redhat.com>