tests: Test extents with qemu full backup flow
Modify the full backup tests to iterate over the image data using
extents. This verifies that qemu return the correct extents, and that
extents works also with qcow2 chain.

The single image test is using also raw format, so we must test it on
filesystem supporting sparseness. Change the test to use user storage.
A nice side affect is testing full backup with both xfs and ext4 and
block size of 512 and 4096, and we can use more real image sizes.

The qcow2 chain test is using only qcow2 format, so we can test it using
tmpdir. We need to add some directory backends to use user storage for
this test.

Change-Id: If19ffc6bd41290c8bc10444039a73539046892bb
Signed-off-by: Nir Soffer <nsoffer@redhat.com>