tests: Mark full_backup raw tests as xfail on Jenkins
This tests fail randomly when trying to right unaligned data on XFS 4k
filesystem:

    ovirt_imageio_common.nbd.ReplyError: [Error 22] bytearray(b'writing to file failed')

I think this is the known issue with 4k block size detection on XFS in
old qemu. This issue was fixed in RHEL 8.1 AV, but we don't have the
advanced virt stream on CentOS yet.

The reason this failure is random is that creating 4K
filesystem on Jeknins fails randomly, so this failure pops up only in
the rare case when we could create 4k filesystem.

Mark this test as xfail on CentOS 8.0 and Jenkins.

Change-Id: If25eafa92627c7e9b072e45f1c868ec44d3dae15
Signed-off-by: Nir Soffer <nsoffer@redhat.com>