tests: Use node names instead of drive name
During backup, a new filter node is placed at the root of the block node
graph. Using "drive0" resolve to the new root node, which does support
adding bitmaps.

The right way to address nodes in the block graph is to use the
"node-name" of the node, which never changes. Set a node name when
starting the vm ("file0") and use it when manipulating the node graph
during backup.

For more info on this see:
https://lists.nongnu.org/archive/html/qemu-block/2019-12/msg00191.html

Change-Id: Ia49774e66578e8a4c6e7e98edb24e03fc80fbe42
Signed-off-by: Nir Soffer <nsoffer@redhat.com>