tests: Prepare for running guest in backup tests
For testing backup we need to run a real guest and do writes inside the
guest.

To control the guest, we use "-serial stdio" to connect qemu stdin and
stdout pipes to the guest serial console. We send commands via stdin,
and get command output and error from stdout.

To make it simple to communicate with the geust, qemu.run() yileds now a
Guest proxy object, using the stdin and stdout to communicate with the
guest.

Finally to speed up guest startup, we use "-nodefaults" and "-net none"
to get a bare machine that boot faster.

Change-Id: Ibe591e5a618110a70f4d6406f94e55e52f27458a
Signed-off-by: Nir Soffer <nsoffer@redhat.com>