conftest: Add base_image fixture
Add base_image session fixture, creating a cirros base image for testing
real guests. The image is created once per session in if it does not
exist yet. On developer environment, it will probably run once. In CI
environment, it will run once per build.

The image is created using virt-builder, so we require now the
libguestfs-tools-c package. This package has quite lot of dependencies,
but the convenience of being able to create images on-the-fly worth it.

This image boot extremely fast. Starting a guest and logging in via the
serial console takes 1.2 seconds.

Creating the base image takes less than 8 seconds on my laptop. On
Travis it takes 41 seconds since we don't have hardware acceleration.

Change-Id: Ibfc9387262dd4f56cfce116097a625d79532b38b
Signed-off-by: Nir Soffer <nsoffer@redhat.com>