tests: Add full backup test with real guest
This is a rather slow test but it does a lot:
- Create a cirros base image if needed
- Create an overlay on the base image to ensure test isolation
- Start a VM from the overlay
- Create a file in the guest
- Starts a full backup with a checkpoint
- Create another file in the guest
- Copy backup data using qemu-img convert
- Stop the backup
- Create an overlay on the backup image
- Start another VM from the overlay
- Check that the file created before the backup exist and the file
  created during the backup does not.

This tests only our backup testing infrastructure; we need to add a new
test using our nbd client or replace qemu-img with the nbd client.

On may laptop, this test takes 2.8 seconds. On Travis it takes 17.6
seconds since we don't have hardware acceleration. This test always
timeout on Jenkins so it is marked as xfail there.

Change-Id: Ia62177d52390bcf8fdefd31671987aa5cf5e15d8
Signed-off-by: Nir Soffer <nsoffer@redhat.com>