nbdutil: Move nbd-client.copy() to nbdutil.copy()
Copying exports is useful for testing backup with real guest. Lets move
it into the nhdutil module.

The new version has one difference, we split big zero extents in the
reader instead of the writer. This is more consistent with the way we
deal with data requests, and can give better concurrency if we add more
than one writer.

Change-Id: I614b5b1cc403962b3194b487a8b99363508d6b21
Signed-off-by: Nir Soffer <nsoffer@redhat.com>