nbdutil: Fix flushing after failures
If the reader failed and stopped the writer, the writer was trying to
flush, which is not needed.

Add "op" field to Request, so the reader sends flush request only on
successful completion.

This will also allow sending flush request during the copy, if we want
to support continue of failed copies from last successful flush.

Change-Id: I9611ce33beca400b5bd2255cafae1f48dc563662
Signed-off-by: Nir Soffer <nsoffer@redhat.com>