tests: Increase qemu shutdown timeout in backup tests
When we create a checkpoint during backup, we must let qemu shutdown
cleanly so it will write the bitmap to the disk.

Previously we killed qemu if it did not terminate in 1 second. This
works most of the time but on Jenkins it can fail randomly like this:

15:43:55 DEBUG   (MainThread) [qemu] Terminating qemu gracefully
15:43:56 WARNING (MainThread) [qemu] Timeout terminating qemu - killing it
15:43:56 DEBUG   (MainThread) [qemu] qemu terminated with exit code -9

When we kill qemu, it cannot write the dirty bitmap, and the test will
fail:

    def list_bitmaps(image):
        info = qemu_img.info(image)
>       return info["format-specific"]["data"]["bitmaps"]
E       KeyError: 'bitmaps'

Add shutdown_timeout argument to qemu.run() and increase the timeout for
the backup tests.

Change-Id: Ibc54de3c9f5d74fc2b4ee09252a576dc8f1810fa
Signed-off-by: Nir Soffer <nsoffer@redhat.com>