nbd: Discover dirty bitmap name during negotiation
Qemu expose the dirty bitmap for an export using the actual bitmap name,
which can be something like "backup-libvirt-42-storage". Looks like the
best way to handle this is to use NBD_OPT_LIST_META_CONTEXT to list all
dirty bitmaps using the empty bitmap name "qemu:dirty-bitmap:".

If the server exports a dirty bitmap we will try to use it, and expose
the value to clients in self.dirty_bitmap, so they can extract the dirty
extents info from the extents() dict.

For more info see the discussion in qemu-block mailing list:
https://lists.nongnu.org/archive/html/qemu-block/2019-12/msg00408.html

Change-Id: Iaebe0e9752fe230a80b9c837054de98483282145
Signed-off-by: Nir Soffer <nsoffer@redhat.com>