tests: Add incremental backup test
Add a test performing full backup with a checkpoint, and then
incremental backup since the checkpoint.

To verify the backup, we rebase the incremental backup disk on top of
the full backup, and use our verify backup helper that run a vm from the
backup disk, and check that files created before the backup exists.

Because we copy only dirty extents, the incremental backup disk is much
smaller:

$ ls -lhs /var/tmp/ovirt-imageio-common/test_incremental_backup_guest0/
 21M -rw-r--r--. 1 nsoffer nsoffer  21M Dec 20 01:17 full-backup.qcow2
1.6M -rw-r--r--. 1 nsoffer nsoffer 1.7M Dec 20 01:17 incr-backup.qcow2

Change-Id: If6f83d382203474fcd74410546a5dcb070dede33
Signed-off-by: Nir Soffer <nsoffer@redhat.com>