daemon: Add ticket "dirty" property
During incremental backup, the NBD server exports a dirty bitmap with
the dirty extents of the image. The NBD client need to locate and
activate this bitmap to allow getting dirty extents. Since NBD client is
created on the first request accessing the image, we must configure the
client using the ticket, in the same way we configure the client to
treat the image as sparse.

Engine should set this property during incremental backup.

Change-Id: I32823b6c7fbab47a07651384bfb27019d4bb4fef
Signed-off-by: Nir Soffer <nsoffer@redhat.com>