backends: Support dirty extents
Backends accept now a "context" argument to extents(). The default value
is "zero", yielding image.ZeroExtent. If called with "dirty", provide
image.DirtyExtent.

The file and memory backends do not support dirty extents, and users are
unlikely to try to use dirty extents when downloading images, so they
raise a new UnsupportedOperation error.

The NBD backend support both zero and dirty extents, but the NBD export
may not support them. If zero extents are not supported, we safely fake
the response in the same way we do for other backends.

If dirty extents are not supported, we raise UnsupportedOperation
because the caller can still use zero extents for more efficient
download. This may happen during incremental backup we lost all the
image checkpoints after unclean shutdown or storage operation that
does not support keeping bitmaps yet.

Change-Id: I381c83833f10a981d71f685f5ec5652b33fb3749
Signed-off-by: Nir Soffer <nsoffer@redhat.com>