examples: Update for new client implementation
Add -f, --format argument to http-client upload, since now we support
download to qcow2 files. qcow2 is also the default since it supports
downloading incremental backups.

Change --buffer-size default to use client.BUFFER_SIZE, so we never need
to change it again. This option is mainly for developers tuning the
default values.

In upload, do not set the progress size to image actual size, as it is
set in the client to the image virtual size.

Change-Id: I72fb028d296d68cc9232d596e7d0a600b8d00e27
Signed-off-by: Nir Soffer <nsoffer@redhat.com>