file: Fix unsafe block size detection
On file backend we detected the block size using a read-write loop. This
allows detection of the block size in all cases, but unsafe when using
multiple connections.

We have 3 issues:
- Data corruption - the write and truncate calls from one backend can
  overwrite real writes from other backends.
- Detection may fail with RuntimeError because of unexpected read size
  after another backend truncated the file during the check.
- When downloading snapshots, writing to read only images used by qemu
  is not a good idea.

We don't use multiple connections for uploads yet, but backup vendors
may try to use that to speed up restores.

Change to use the same way qemu detects block size:

1. Try to read 1 byte
2. If succeeded, we cannot detect the block size, fallback to 4096
3. Try to read 512 bytes
4. If succeeded, use 512
5. Try to read 4096
6. If succeeded, use 4096

With this change, we will use block size of 4096 bytes with:
- Empty raw-sparse image on XFS or Gluster/XFS with 512 bytes sector size
- NFS, since there is no alignment requirements

To consider this limit in the tests, storage.Backend has now a
can_detect_sector_size attribute.

Two useless tests for opening a backend in write-only mode ("w") were
removed.  This is not a real use case and we should never truncate
images.

To make our life more interesting, turns out that on Travis block size
detection works even with XFS, and in Jeknins it works in the Fedora 30
builds, but not in el8 builds. This seems to depend on the underlying
kernel, so the tests ensure that the detected block size is same or
bigger than the actual block size.

Change-Id: I217d30f353344bb0d7256cdcfffeafd7bf0052f5
Signed-off-by: Nir Soffer <nsoffer@redhat.com>