backends: memory: Remove unused write_to()
The memory backend implemented the older write_to() interface that did
not accept a buffer, but it was never used. Remove it instead of
updating it to accept a buffer.

write_to() is an optimization for backends that need to control the read
loop. The http backend use this to perform single request per call, but
the memory backend does not have any reason to use the optimization.

Change-Id: I8af791a1a894aaab0d4dc6e3ed4a5337b361a2dc
Signed-off-by: Nir Soffer <nsoffer@redhat.com>