backends: nbd: Remove unused write_to()
The nbd backend implemented the old write_to() interface (that did not
accept a buffer) so we could avoid the unwanted copy in readinto() from
the backend buffer to operation buffer, but it was never used.

In python 3 the aligned buffer used by the ops module can be used with
memoryview, so we can pass it to the nbd client, and write_to() is not
needed.

write_to() should be implemented only if the backend need to control the
read loop. Otherwise generic read loop should be used.

The _buffer instance variable is unused now, so it was removed as well.

Change-Id: Ib9f02470fd279c2be70f7b0426dfb9d496197124
Signed-off-by: Nir Soffer <nsoffer@redhat.com>