backends.file: Remove buffer for zeroing
The file backend was allocating additional 8 MiB buffer for zeroing
manually, if the underling filesystem does not support fallocate. People
using legacy storage that does not support fallocate do not care about
performance, so we should not complicate the code to make it faster.

Replace the buffer with temporary buffer allocated when zeroing. This
make the code simpler and remove the need to pass the buffer size to the
backend.

Since we require now python 3 for the daemon, replace usage of
compat.bufview with memoryview, and skip tests using file backend on
python 2.

Change-Id: If1dd74eececf599bd0e01bad2606a6bc94c55d7c
Signed-off-by: Nir Soffer <nsoffer@redhat.com>