tests: Skip tests using nbd module on python 2
We need to use memoryview for improving performance of the nbd module,
which is our core infrastructure for upload and download. The nbd module
is not used by the proxy, which is the only component that need to be
compatible with python 2, so this is not an issue for users.

Mark all the tests for the nbd module or using the nbd module in the
tests infrastructure as @requires_python3.

Change-id: I399fa46c948b93bf2e3ab00b66e6ea029d884f94
Signed-off-by: Nir Soffer <nsoffer@redhat.com>