qemu_nbd: Use --argument=value form
qemu-nbd supports both "--argument value" and "--argument=value". The
second form is more readable but we could not use it before since in
python 2 we could not pass unicode values to qemu-nbd. Since we require
now python 3, switch to the more readable form, already used in most of
this module.

Change-Id: Ib26efffb16878ed715ec61c18f3970394e037cff
Signed-off-by: Nir Soffer <nsoffer@redhat.com>