qemu_nbd: Make more arguments configurable
For the tests we could hard code lot of values that may not be good
choice for the client, running on unknown environment, or used in
specific configuration for performance testing. Make the arguments
configurable.

The defaults were not changed, using the same configuration used in the
tests, and on a hypervisor:

    --cache=none --aio=native --discard=unmap

This configuration usually show best performance and most predictable
results since we avoid the host buffer cache. However when testing
maximum throughput using /dev/shm or /dev/null, direct I/O is not
supported.

Setting the new options to None will disable these options, using
qemu-nbd defaults that are more compatible with unknown environments.

Change-Id: I0b28bf1fc3943d510f195bc926135bbf4b99875f
Signed-off-by: Nir Soffer <nsoffer@redhat.com>