qemu_nbd: Avoid qemu-nbd ready check warnings
Using nbdutil.wait_for_socket() cause this warning on every run:

    qemu-nbd: option negotiation failed: Failed to read flags: Unable to
    read from socket: Connection reset by peer

qemu-nbd is warning that a client negotiation has failed. This is
expected since we test that qemu-nbd is ready by trying to connect the
socket.

In the tests this warning is typically hidden, and we see it only when a
test fails, so I did not bother to handle this. Now that we want to use
qemu-nbd in the client, we must avoid this warning.

Configure qemu-nbd child process to collect child process stderr, and
log collected errors if qemu-nbd terminated abnormally.

Change-Id: I0599ff704dbf4251d4c17cd5678cee525a573821
Signed-off-by: Nir Soffer <nsoffer@redhat.com>