common: Introduce the io module
The io module implements io operations for backends. The first operation
is io.copy(), copying source backup to destination backup.

The copy uses source backend extents to copy only the data extents from
source backend to destination backend.

By default, io.copy() zero unallocated areas on the destination. When
copying to new image on file based storage this is not needed; calling
with zero=False will skip zeroing.

If the destination implements read_from(), it is used to copy data
extents from source. Otherwise if the source implements write_to(), it
is used to write data to the destination. Otherwise fallback to generic
copy that works with backends implementing only readinto() and write().

Finally, large extents are splitted to smaller ones for more frequent
progress updates when dealing with very large zero or data extents that
may take lot of time to process.

Change-Id: I0ec38cf5ee330c6cfcd907ddab390d7b8acc8ef7
Signed-off-by: Nir Soffer <nsoffer@redhat.com>