client: Reimplement client using http and nbd backends
Now that we have http and nbd backends, we can implement download and
upload using http backend for talking with imageio, and nbd backend for
talking with qemu-nbd.

We use this pipeline for download:

    imageio -> http-backend -> nbd-backend -> qemu-nbd

And this for upload:

    qemu-nbd -> nbd-backend -> http-backend -> imageio

With this we can support any image format supported by qemu as input and
output, for example, download raw disk to qcow2 image, or upload qcow2
image to raw disk.

Client test for splitting big zero request was removed since the client
does not handle this now. This should be tested in the io module tests.

Change-Id: Ie1eb493e2dd7a15af8c33fdc0e800ec4d2bd6c92
Signed-off-by: Nir Soffer <nsoffer@redhat.com>