backends: memory: Support external backing
Change memory backend to accept external bytearray() instead of copying
the data to internal io.BytesIO(). This makes it possible to create
multiple readers sharing the same backing.

Since the backing file may be shared, the backend does not not own it
and does not close it. Instead the backend maintain closed state to
behave like other backends.

The memory backend maintains its own position like the nbd and http
backends. Seeking in one backend does not affect other backends.

Change-Id: Icb82df381bacadec93a7f610cfdba99f80a005ed
Signed-off-by: Nir Soffer <nsoffer@redhat.com>