tests: copy test utils into common package
Most of the tests in the in daemon package use testutils. As we want
to move modules into common package gradually with minimal possible
changes to make reviews more easy, we need this file in both packages
temporarly. Copy testutils module into common package and move its
tests there as well.

This is not a nice solution, but other solutions, like hacking import
path, would be even more ugly.

Change-Id: Ifdaa1db7834833a85f531ea29f093b084c60084b
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>