daemon: use version from common package
Version module is the only one remaining in daemon package (besides
config files and script). Version is also in common module. Switch
to version from common modules so that we can remove directories
other than that used for building daemon package.

As at the point of building the package ovirt-imageio-common package
is not installed in in the environment, we need to adjust system path.
This is a temporal hack as we will remove daemon package in the future.

Change-Id: Ie8cfc035e5385aa9068c47ded39eb0d5085abde1
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>