daemon: remove src and test directories
As we removed all the modules and test in previous patches, src and test
directories can be removed now. We can also remove tox.ini as it's
useless now as we haven't any tests and calling check target in
Makefile. However, as during the tests we call check target explicitely,
we need to provide empty check target, otherwise build would fail.
We can also remove TODO files as it's not useful any more.

At this point we really keep in daemon package only stuff needed for
building the package.

Change-Id: Ic8fc90baa96ba77ed235c500afbb0148c0f7767a
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>