tests: Don't use global config in http helper
We used to import config in http helper, and use
config.images.{host,port} for creating a connection. This will not work
when we want to test proxy and daemon in the same test, and proxy is
implemented as a daemon with proxy ticket:

    proxy -> daemon -> storage

In this setup we need to start 2 daemons, each with different
configuration, so we cannot use a global config in the tests.

The first step to fix this is to stop using the global config in the
http module. Instead, add http.Client() class, accepting a config object
and use the config to create the connection. The tests still use a
global config to create the client.

For consistency, create also a http.UnixClient() class, so tests using
config.{images,tickets}.socket can use the same code as tests using
config.images.{host,port}. This is important for extents_tests running
all tests with both http and local services.

Change-Id: I42643d9b43a3f157548551bca674a49b289dd75a
Signed-off-by: Nir Soffer <nsoffer@redhat.com>