services: Extract services module
The server module included both the server state and the services
classes. Extract services module for the services classes.

This services are already used by some of the tests, while other tests
use the server module (server.start(), server.stop()). Because we need
to start 2 servers, we will convert all tests to use the services
classes instead of the server module.

Change-Id: I82853da2d90dbed9b2152858fd8d2755ebe26fdb
Signed-off-by: Nir Soffer <nsoffer@redhat.com>