config: Introduce config.load()
We use configloader.load() to modify the config module, and then we pass
the config module object to the server. The server pass the config
object to the reset of the application, so the application does not use
the config as global.

However many tests did the shortcut of importing the config and loading
test config for starting a server. This prevent testing 2 daemons, one
serving as proxy, and the other as daemon.

Introduce config.Config() class, creating an independent config object
inheriting all the defaults without modifying them, by creating an
instance of all the sections in the config object.

Introduce a config.load() function for loading new configuration object
from storage.

Creating a configuration is easier now:

    cfg = config.load(["/path/to/config"])

Change-Id: I85d5ead538c1958a0261607d9c1c78bd4d807ca4
Signed-off-by: Nir Soffer <nsoffer@redhat.com>