tests: Stop using global config
The server and tests use now private config object without modifying the
config module.

Tests starting and stopping the server use now a cfg fixture loading
configuration and running the server while module tests are running.
This is little ugly but keeps the change smaller. This will be improved
when we convert all tests to run services using fixtures.

Change-Id: I729cb1fefa22e60bb76d3cf1e372de9c0e797b4c
Signed-off-by: Nir Soffer <nsoffer@redhat.com>