auth: Prepare for using private auth
The auth module is used as a singleton, which works in the real
application. However this prevents testing proxy and daemon services in
the same tests, since both of them will share the same auth._tickets
map.

Prepare for using private auth by importing the auth module in the
server and passing it to the services. The services pass the auth object
to the request handlers.

Change-Id: I30752eb4388b4799b08978d3700e656404c4ffc2
Signed-off-by: Nir Soffer <nsoffer@redhat.com>