testutil: Use unique ticket uuid by default
We used the same uuid for all tickets by default. This is bad default
since tickets have unique UUID in almost all cases, and can lead to test
failures when ticket added by one test cause a later test to fail.

We have only one case when we need 2 tickets with same uuid, when we
test proxy and daemon flow, and we need proxy ticket and daemon ticket
with the same UUID.

Change-Id: If04c7a00d419d21ee081bf66480d1f5f90a71e39
Signed-off-by: Nir Soffer <nsoffer@redhat.com>