auth: Stop using global auth state
Instead of keeping the tickets in the auto._tickets, move the state into
new auth.Authoerizer() class. The server creates an authorizer and pass
it to all services.

The tests create authorizer per test module and pass it the server or
service being tested.

All tests were modified to use a uniform env() fixture for creating the
test environment.

We are ready now for creating a test with both proxy and daemon
services, each with different config and authorizer.

Change-Id: Ib5dcedcdf105f45c9391b429460a38ca42970172
Signed-off-by: Nir Soffer <nsoffer@redhat.com>