setup: use own user and group
We want to decouple imageio from vdsm service. As the first step,
imageio need to use its own user and group under which it will run.

Create new user ovirtimg and it's group with the same name. Use this
user and group to run imageio service.

At this poing imageio still depends on vdsm, as it still uses
its certificates.

Change-Id: Ice798bfa225549e3cafa80b1d3f5eb9b06f32839
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>