backends: Pass config to backend
The http backend needs cafile argument for verifying server
certificates. When used in the client, the cafile is specified by the
user. When used in the proxy, the cafile come from the proxy config.

Pass the handlers config to backends.get() to it can pass the cafile to
the http backend.

Change-Id: I1389b05b7019ef10486e4a051a65028836e8898d
Signed-off-by: Nir Soffer <nsoffer@redhat.com>