control: make socket writable for whole group
Imageio now runs under its own user, but other services it integrates
with, e.g. vdsm, still need socket created by control service to be
writable also for them. By default, socket is created in mode 0o755
and thus even if the service is part of ovirtimg group, it still fails
to write to imageio's control socket. Make control socket writable for
whole group and use mode 0o660 as there's no need to make it readable
for the rest of the users.

Change-Id: I8a0eb517f0a23f96ed32f8fc52215b3de0206f30
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>