tests: Add failing test for short read
Unlike all the other backends, the http backend does not support short
reads. When trying to read after the end of the image, it sends invalid
content range to the server.

Fix the test http server to behave like the real daemon when getting
invalid content range, and add a failing test reproducing this issue.

Add failing tests for short reads, and reading at or after the end of
the image.

Change-Id: I0f8bde00f57be607f91e3560defb5da9fb65d5a1
Signed-off-by: Nir Soffer <nsoffer@redhat.com>