backends: http: Support short reads
Support short reads like all other backends. If readinto() buffer is
larger than the available data, fill the buffer with the available data
and return the number of bytes read.

For the http backend this means clipping the length of the request to
the size of the image instead of sending invalid content-range to the
server.

The special cases of zero length request or request after the end of the
image are handled explicitly, since HTTP Range does not allow zero
length requests[1].

This bug did not affect the client since it never reads behind the end
of the image, but the ops.Send() depend on short reads to handle
unaligned requests.

With this change the proxy supports unaligned downloads.

[1] https://tools.ietf.org/html/rfc7233#section-2.1

Change-Id: I8dae7dd7042a58c51523662e970916f562409e70
Signed-off-by: Nir Soffer <nsoffer@redhat.com>