backends: nbd: Fix readinto() at and after end
The NBD protocol suggest to avoid zero length read request:

    The client SHOULD NOT request a read length of 0; the behavior of a
    server on such a request is unspecified although the server SHOULD
    NOT disconnect.

In this case we sent a zero length request. qemu is handling the query
gracefully, but we should not do this.

Requests after the end of the image are not allowed. I could not find
anything in the spec about that, but it makes sense since block devices
do not allow this. In this case we send invalid request.

Fix both cases by returning 0 without sending any request.

Change-Id: I402d556b864096fe7b8ce064a02b529ced590a0a
Signed-off-by: Nir Soffer <nsoffer@redhat.com>