automation: Disable el7 builds
We still build on el7 for the ovirt-imageio-proxy package, but these
builds fail now with:

    https://cbs.centos.org/repos/virt7-ovirt-43-testing/ppc64le/os/repodata/repomd.xml:
    [Errno 14] HTTPS Error 404 - Not Found

Since we are going to drop the proxy very soon, disable these builds
now.

We sill have el7 build in Travis, which is not affected by the CentOS
repo since we have all the dependencies in the centos-7 container image.

Removing el7 files under automation/ and docker/ will be done later.

Change-Id: I05facf84b821eff1fe6280ccbc986721082cb6a5
Signed-off-by: Nir Soffer <nsoffer@redhat.com>