daemon: Support CORS requests
When accessing the daemon via engine web admin, the browser is sending a
preflight OPTIONS request to check if CORS protocol is supported, and
the server allow certain headers and methods.

When sending GET or PUT request, the browser specify the Origin header
and expects to get a Access-Control-Allow-Origin header matching the
requested origin.

This was available before in ovirt_imageio_proxy/http_helper.py. This
version return similar response, with these differences:

- Return Access-Control-Allow-* only if the client specified the
  matching Access-Control-Request-* header to minimize response header
  overhead.

- Allow any header so we don't have to modify this code again.

- Don't send Access-Control-Expose-Headers since we don't use special
  headers which are not safelisted[1] by default.

- Increase Access-Control-Max-Age to avoid unneeded repeated preflight
  requests every 300 seconds.

[1] https://developer.mozilla.org/en-US/docs/Glossary/CORS-safelisted_response_header

Change-Id: I78f7a2005b63bc32853a99a1d1f38d3d38a32bc2
Signed-off-by: Nir Soffer <nsoffer@redhat.com>