spec: Try to stop old service on upgrade
When upgrading from ovirt-imageio-daemon < 2, the old
ovirt-imageio-daemon service is left active after the upgrade. This
causes the new ovirt-imageio service to fail with:

    OSError: [Errno 98] Address already in use

Add a %posttrans secriptlet ensuring that at the end of the transaction
ovirt-imageio-daemon.service is stopped.

Vdsm %posttrans runs after ovirt-imageio scriptlet and start vdsm if
needed. When vdsm starts it will start the new ovirt-imageio service
if needed.

Change-Id: If82ea81a605e2bb657e456d6478598807b3aefbb
Signed-off-by: Nir Soffer <nsoffer@redhat.com>