tests: unify test clients
In the future we would like to provide support for communication with
control service over TCP. To be able to test it easily, we need to unify
all the client to be constructed in the same way, so we can easily
parametrize the fixures and get either unix socket client or HTTP
client.

We use unix socket client in local and tickets tests. Create
dedicated classes for local service client and control service client,
both accepting deaemon config as a constructor parameter.

Change-Id: I510f16fea76c49ca2909d331f30ddd59e0eadafd
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>