tests: unify http and unix clients
In previous patch we refactored HTTP client into a general client, which
now accepts a connection as a constructor parameter. We leverage this to
simplify also UnixClient to inherit from HTTPClient and thus simplify
the code and also some tests, which now can use get, put and patch
methods of the client directly.

Change-Id: Id8d175fb4fd6cc0495ed816dba0663ff5b5b2d5f
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>