backends: http: Re-raise GET /extents errors
Like errors in GET /images/ticket-id, we also reraise errors in GET
/images/ticket-id/extents.

While the error code and message from the daemon is passed correctly to
the client, internal error on the daemon side is logged as internal
error in proxy side, which is not desired. More work is needed to
to avoid logging traceback for remote errors.

Change-Id: I6ee2ec354dc65bb9bb74a6704ab6f7c73719c60e
Signed-off-by: Nir Soffer <nsoffer@redhat.com>