ops: Close subviews on errors
When writing to a subview fail, the request fails with internal error
instead of the http.Error raised from the backend:

Traceback (most recent call last):
  File "/home/nsoffer/src/ovirt-imageio/daemon/ovirt_imageio/http.py", line 694, in __call__
    self.dispatch(req, resp)
  File "/home/nsoffer/src/ovirt-imageio/daemon/ovirt_imageio/http.py", line 739, in dispatch
    return method(req, resp, *match.groups())
  File "/home/nsoffer/src/ovirt-imageio/daemon/ovirt_imageio/cors.py", line 84, in wrapper
    return func(self, req, resp, *args)
  File "/home/nsoffer/src/ovirt-imageio/daemon/ovirt_imageio/images.py", line 69, in put
    ticket.run(op)
  File "/home/nsoffer/src/ovirt-imageio/daemon/ovirt_imageio/auth.py", line 120, in run
    operation.run()
  File "/home/nsoffer/src/ovirt-imageio/daemon/ovirt_imageio/ops.py", line 62, in run
    self._run()
  File "/home/nsoffer/src/ovirt-imageio/daemon/ovirt_imageio/ops.py", line 145, in _run
    except EOF:
  File "/usr/lib64/python3.7/contextlib.py", line 298, in __exit__
    self.thing.close()
BufferError: cannot close exported pointers exist

We already use a memoryview as context manager to close the view, but we
we write to a subview, and it seems that this subview need to be closed
as well.

Change both read and write loops to close each subview.

Change-Id: Ia50c02001687f8cf5c6502b42b038fe46fa27cdb
Signed-off-by: Nir Soffer <nsoffer@redhat.com>