build: Inject release suffix when generating spec
The release suffix (.timestamp.githush) was injected when running
rpmbuild in "make rpm". This works in jenkins since we build the rpms
using "make rpm", but in other build systems invoke rpmbuild directly so
we always get empty release suffix. This is fine for release build, but
does not work for copr builds.

Change the makefile and spec to inject the release suffix when
generating the spec.

When building from tag, release suffix is empty, so this does not affect
release source rpm.

Change-Id: Iaba822747629230415517807f7e2267fd2585157
Signed-off-by: Nir Soffer <nsoffer@redhat.com>