tests: add test for dropping privileges
Add tests which tests that we drop privilegs when the daemon is started
under root. We want to drop privilegs into specified user in such case
and also make sure that run directory, control socket and log file are
owned by this user. We don't want to drop privilegs if this is
explicitely configured in daemon config. If the daemon is run under
non-root user, dropping the privileges should be skipped and daemon
should continue running under this user.

Change-Id: I3f154bd8b746cb697abe6fee60b6c5adbaeb9d4e
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>