util: add general socket utilities
nbdutil.wait_for_socket() is general function who to wait for the socket
which can be used e.g. in the test running imageio daemon to wait for
it to start. Move this function to sockutil module to make it
re-useable. Move there alse UnixAddress and TCPAddress which this
function needs and keep in nbd module their subclasses which provide
nbd-related functionality.

Change-Id: If30d47f8da690eb948d203f62fba8b11c9d4073c
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>