stats: Support recording number of bytes
Support collecting number of bytes per every operation. This is more
tricky then expected, since we know the number of bytes only when I/O
completes, so we cannot pass it to start() or run().

To solve this, Clock.start() and Clock.run() return a Stats instance,
and the caller need to update the number of bytes manually.

Clock.__repr__ shows now the total number of bytes and total rate
for this operation formatted in human friendly manner.

NullClock use special NullStat instance that drop the bytes values, so
code does not need to care if stats collections is enabled.

The typical use case is:

    with clock.run("name") as stats:
        n = backend.write(buf)
        stats.bytes += n

When the caller cannot use Clock.run(), it should keep the stats
instance and update it before calling stop.

    stats = clock.start("name")
    ...
    stats.bytes += 4096
    clock.stop("name")

Here is output from the benchmark, showing the overhead of measuring
50,000 requests:

    [connection 1 ops, 0.366829 s] [request 50000 ops, 0.334896 s]
    [read 50000 ops, 0.030955 s, 97.66 GiB, 3.08 TiB/s]
    [write 50000 ops, 0.031029 s, 97.66 GiB, 3.07 TiB/s]

Change-Id: I7dc223ec077284cacae1ceebb343b044b198cc3f
Signed-off-by: Nir Soffer <nsoffer@redhat.com>