stats: Refine error handling semantics
When using Clock.start(), count only successful operations - when
Clock.stop() is called. Previously we counted the operation when they
were started.

If an exception is raised in run() before updating number of bytes, the
operation is not counted. Previously we always counted the operation,
even if the user code raised.

To allow same semantics when using Clock.start(), add Clock.abort().
Calling Clock.abort() will stop the operation without counting it.

Here is an example to make this more clear:

    with clock.run("write") as s:
        n = write(buf)
        s.bytes += n

    s = clock.start("write")
    try:
        n = write(buf)
        s.bytes += n
    except:
        clock.abort("write")
        raise

In both cases the write operation is counted only if write() succeeded.
The time spent waiting for write() is always counted.

With these semantics we can implement retries:

    for retry in range(1, max_retries):
        try:
            with clock.run("write") as s:
                n = write(buf)
                s.bytes += n
                return n
        except:
            if retry == max_retries:
                raise

Only one "write" will be counted, but the time will include the time of
the failed write() calls.

Change-Id: I69e276521d141dea4afa50ac65657a6338ac2354
Signed-off-by: Nir Soffer <nsoffer@redhat.com>